[roll-out-web] Don't assume nginx and/or web were booted before (#6207)
The implicit assumptions in `bin/server/roll-out-web.sh` that nginx
and/or web will have been booted before that script is executed cause
`GIT_REV=$(git log main --format=format:%H | head -1)
bin/server/deploy.sh` to fail on `main` when run on a fresh server
(which I discovered when running it on a fresh Hetzner server just
recently).

This change modifies the `bin/server/roll-out-web.sh` script so that it
will work even if `nginx` and/or `web` were _not_ booted before that
script is executed.

+++ bin/server/roll-out-web.sh
@@ -73,13 +73,21 @@ while [ $attempt -le $max_attempts ] ; do
   fi
 done
 
-# Start routing requests to the new container instead of the old.
-reload_nginx
+if docker compose ps -q nginx | grep -q .; then
+  # If NGINX is already running, reload to have it start routing requests to the
+  # new container instead of the old.
+  reload_nginx
+else
+  # If NGINX is not yet running, boot it up.
+  docker compose up --detach --remove-orphans nginx
+fi
 
 # Remove the old container.
-echo_iso8601 "Removing the old container."
-docker stop "$old_container_id"
-docker rm "$old_container_id"
+if [[ -n "$old_container_id" ]]; then
+  echo_iso8601 "Removing the old container."
+  docker stop "$old_container_id"
+  docker rm "$old_container_id"
+fi
 
 scale_web 1
